+++ tests/testthat/test-sf-from-raster.R
@@ -12,10 +12,17 @@ test_that("creation of polygons from raster works", {
   expect_that(qm_rasterToPolygons(r), is_a("sf"))
   expect_that(qm_rasterToPolygons(r), is_a("sf"))
   
-  expect_that(qm_rasterToPolygons_sp(b), is_a("SpatialPolygonsDataFrame"))
-  expect_that(qm_rasterToPolygons_sp(b), is_a("SpatialPolygonsDataFrame"))
-  expect_that(qm_rasterToPolygons_sp(r), is_a("SpatialPolygonsDataFrame"))
-  expect_that(qm_rasterToPolygons_sp(r), is_a("SpatialPolygonsDataFrame"))
+  # expect_that(qm_rasterToPolygons_sp(b), is_a("SpatialPolygonsDataFrame"))
+  # expect_that(qm_rasterToPolygons_sp(b), is_a("SpatialPolygonsDataFrame"))
+  # expect_that(qm_rasterToPolygons_sp(r), is_a("SpatialPolygonsDataFrame"))
+  # expect_that(qm_rasterToPolygons_sp(r), is_a("SpatialPolygonsDataFrame"))
+  # 
   
-  
+})
+
+d <- raster::rasterize(lux, raster::raster(lux, res = 0.01))
+test_that("we can also qm_raster here", {
+  qm_rasterToPolygons(d) %>% 
+    expect_named(c("ID_1", "NAME_1", "ID_2", "NAME_2", "AREA", "geometry")) %>% 
+    expect_s3_class("sf")
 })