+++ NEWS.md
@@ -1,3 +1,9 @@
+# spex dev
+
+* added S3 generic polygonize, alias of qm_rasterToPolygons
+
+* removed dependence on sf
+
 # spex 0.2.0
 
 * Added function `qm_rasterToPolygons`.